timer_handlerにポモドーロタイマーテーブルを操作する機能の移転
